Refactor IF/ELSE running and result model. Part of #3749.
Earlier IF/ELSE structures were represented so that `If` objects
represented IF branches and they has `orelse` attribute containing
next ELSE IF or ELSE branch. ELSE (IF) branches were also represented
by `If` objects which, again, had `orelse`. Due to this nested
structure branches needed to have both the normal `status` attribute
(tells the overall status, including `orelse` status) and a separate
`branch_status` attribute (tells the status of this particular
branch). The nested structure was also used in output.xml:

<if type='if' condition='$x > 0'>
    <if type='else'>
        <kw .../>
        <status status='PASS'/>
    </if>
    <status status='PASS' branchstatus='NOT RUN'/>
</if>

After this commit IF/ELSE structures are represented so that there's a
root `If` object which has `body` containing actual IF/ELSE branches.
This removes the need for `orelse` and `branch_status`, but requires
adding a new model object representing the root. The old `If` class was
renamed to `IfBranch` that now represent a single branch. Same structure
used also in output.xml:

<if>
    <branc type='if' condition='$x > 0'>
        <status status='NOT RUN'/>
    </branch>
    <branch type='else'>
        <kw .../>
        <status status='PASS'/>
    </if>
    <status status='PASS'/>
</if>

There are pros and cons with both ways to represent IF/ELSE
structures, but overall the new approach feels more straightforward
and is probably easier to understand and use for others.

Notice that the parsing model still uses the `orelse` approach. There
it works well and I don't think changing it would be worth the effort.