Fix listeners running kws inside FOR loop
Fixes #4112.

Added tests also for listeners with IF/ELSE. Listeners aren't called
with the root of the IF/ELSE structures so there were no problems.
Listeners are called also with unexecuted IF/ELSE branches. That's
perhaps a bit strange but I consider it a feature rather than a
bug. Listeners can check the type of the executed "keyword" if needed.